ocf_irc: Improve the inspircd service file
The service file provided by Debian is really dated. Add overrides with
some options that upstream has added. I decided this was worthwhile when
the IRC server on dev-flood went down for a couple days due to a SIGSEV,
and was never brought back automatically since Restart= was not set.

https://github.com/inspircd/inspircd/blob/insp3/make/template/inspircd.service